refactor(rql): move SQL frontend dispatch (sql.rs) into reddb-io-rql
The SQL frontend command surface (SqlStatement/FrontendStatement/
SqlCommand + into_query_expr lowering) and the bulk of the parser
dispatch (Parser::parse_frontend_statement, parse_sql_statement,
parse_sql_command, …) lived in storage/query/sql.rs — the last
impl Parser block outside the moved parser dir. Move it into
reddb-io-rql/src/sql.rs so Parser::parse() resolves
parse_frontend_statement in-crate; server storage::query::sql becomes
a re-export shim. Depends only on reddb-io-types; graph stays acyclic.

Carry the source crate's #![allow(dead_code, unused_imports,
unused_variables)] onto reddb-io-rql so the relocation stays a pure
byte-faithful move (matching the reddb-io-types precedent, ADR 0052).

Refs #1103
